Structure deployment information by section
The deployment file was changed to specify the headers of the underlying sections. This will make it easier to control the presentation of this chapter of the report and make it more consistent. Apart from this the server configuration part will be contained within one section in the report and its header specified in the main deployment file.