Change to twosided report
This commit will make the report treat odd and even pages differently and
provide a better experience in printed format. It is only the headers at the top
of the pages that get affected by the change. They are adjusted so that the page
numbers are always presented at the outer edge of a page.